apps/backend-agent/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-core
langchain-community
langchain-ollama
pydantic
python-dotenv
httpx
asyncpg
psycopg2-binary
tavily-python
langchain-openai
beautifulsoup4
langchain-ollama
langchain-google-genai
duckduckgo-search
